Enable native ligatures (w/hack to disable extras)
See https://github.com/hlissner/doom-emacs/issues/5738

I want the native ligatures, but I don't want the extras supported by
prettify-symbols-mode (in particular I prefer not to have the
python-mode conversions of `and` to `∧` -- I admit that symbol
usage is a cool idea, but my brain just can't read it easily). It seems
as if omitting the `+extra` flag in init.el is not disabling those
extras as I would expect, and I'm familiar enough with
prettify-symbols-mode to know that I'm OK disabling it altogether, so I
took the nuclear option to forcefully disable it any time it gets
enabled by anything.

Obviously this is a hack so use this approach at your own risk.